requirements.txt
pymysql
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-dotenv
jose[cryptography]
